Version 5.5 更新库函数, 献祭摄像头成就无线串口, 遥控推车

+++ code/Position_Calculation.h
@@ -15,7 +15,7 @@
 
 
 //===================================================全局变量BEG===================================================
-extern int    Track_Points_NUM;                                // 当前追踪第几个点
+extern int16  Track_Points_NUM;                                // 当前追踪第几个点
 extern double Angle_Error;                                     // 角度误差
 extern float  Fusion_angle;                                    // GPS和IMU互补滤波后的角度
 extern float  Fusion_alpha;                                    // GPS和IMU互补滤波的权重
